requirements.txt
accelerate==1.10.1
albumentations==1.3.1
einops==0.8.1
matplotlib==3.10.7
numpy==2.3.4
openai_clip==1.0.1
opencv_python==4.11.0.86
opencv_python_headless==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
pycocotools==2.0.10
scikit_learn==1.7.2
seaborn==0.13.2
setuptools==78.1.1
tabulate==0.9.0
timm==1.0.21
torch==2.6.0+cu118
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.56.2
